requirements.txt
pytorch
numpy
cv2
spicy